Capture unhandled exceptions
Since there should be only one PACKMAN process with some subcommands, the
signature file '.pid' must be created and deleted correctly.

The unhandled exceptions will cause that file not be deleted, and this
will prevent other PACKMAN processes running. So I added a capture for
those exceptions to ensure that file is deleted.